pyproject.toml
[project]
name = "weather-outfit-advisor"
version = "0.1.0"
description = "ReAct agent for weather-appropriate outfit suggestions using Google ADK"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.14.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "wikipedia>=1.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
